backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Configuration
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# Agent Development Kit
google-adk==1.19.0
litellm>=1.50.0

# LLM
google-generativeai>=0.8.0
openai>=1.50.0
